:hammer: Simplify how the state format regular expression is built
Rather than import each of the compass directions on their own and string
them together by hand (which is fine mostly, there's only 4), do it in a way
where the journey reader doesn't need to have knowledge of how many compass
points there are.

This would allow for easier changes.